Move actions to site-dismissals
