+++ client/blocks/reader-recommended-sites/index.jsx
@@ -14,7 +14,7 @@ import { connect } from 'react-redux';
  */
 import { recordAction, recordTrackWithRailcar, recordTracksRailcarRender } from 'reader/stats';
 import Button from 'components/button';
-import { dismissSite } from 'state/reader/streams/actions';
+import { dismissSite } from 'state/reader/site-dismissals/actions';
 import ConnectedSubscriptionListItem from 'blocks/reader-subscription-list-item/connected';
 
 export class RecommendedSites extends React.PureComponent {
